cloud_functions/ingest_bts_csv/requirements.txt
functions-framework==3.8.1
google-cloud-bigquery>=3.40.0
google-cloud-storage==2.18.2
requests==2.32.3